Trait substrate_state_trie_migration_rpc::StateMigrationApiServer

pub trait StateMigrationApiServer<BlockHash>: Sized + Send + Sync + 'static {
    fn call(&self, at: Option<BlockHash>) -> RpcResult<MigrationStatusResult>;

    fn into_rpc(self) -> RpcModule<Self>
    where
        BlockHash: Send + Sync + 'static + DeserializeOwned,
    { ... }
}

Server trait implementation for the StateMigrationApi RPC API.

Required methods

fn call(&self, at: Option<BlockHash>) -> RpcResult<MigrationStatusResult>

Check current migration state.

This call is performed locally without submitting any transactions. Thus executing this won’t change any state. Nonetheless it is a VERY costy call that should be only exposed to trusted peers.

Provided methods

fn into_rpc(self) -> RpcModule<Self> where
BlockHash: Send + Sync + 'static + DeserializeOwned,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.